+++ sdk/core/azure-core/src/http/winhttp/win_http_transport.cpp
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/core/azure.hpp"
 #include "azure/core/http/http.hpp"
 #include "azure/core/http/winhttp/win_http_client.hpp"
 
@@ -15,8 +14,8 @@ WinHttpTansport::~WinHttpTansport() {}
 
 std::unique_ptr<RawResponse> WinHttpTansport::Send(Context const& context, Request& request)
 {
-  AZURE_UNREFERENCED_PARAMETER(context);
-  AZURE_UNREFERENCED_PARAMETER(request);
+  void(context);
+  void(request);
 
   throw;
 }